+++ refact_webgui/webgui/selfhost_model_resolve.py
@@ -1,7 +1,7 @@
 import json
 
-from self_hosting_machinery.webgui.selfhost_queue import InferenceQueue
-from self_hosting_machinery import env
+from self_hosting_machinery import env  # REFACTORME
+from refact_webgui.webgui.selfhost_queue import InferenceQueue
 
 from typing import Tuple, List
