+++ refact_webgui/webgui/selfhost_plugins.py
@@ -1,4 +1,3 @@
-import json
 from fastapi import APIRouter, Request
