+++ refact_webgui/webgui/selfhost_static.py
