+++ refact_webgui/webgui/static/assets/style.css
