+++ refact_webgui/webgui/static/components/modals/modal-upload-files.js
