+++ refact_webgui/webgui/static/dashboards/dash_prime.js
