+++ refact_webgui/webgui/static/dashboards/dash_users.js
