+++ refact_webgui/webgui/static/dashboards/plots.js
