+++ refact_webgui/webgui/static/dashboards/utils.js
