+++ refact_webgui/webgui/static/error.js
