+++ refact_webgui/webgui/static/favicon.png
