+++ refact_webgui/webgui/static/index.js
