+++ refact_webgui/webgui/static/login.js
