+++ refact_webgui/webgui/static/style.css
