+++ refact_webgui/webgui/static/tab-chat.html
