+++ refact_webgui/webgui/static/tab-chat.js
