+++ refact_webgui/webgui/static/tab-server-logs.html
