+++ refact_webgui/webgui/static/tab-stats.html
