+++ refact_webgui/webgui/static/tab-stats.js
