+++ refact_webgui/webgui/static/tab-upload.js
