+++ refact_webgui/webgui/tab_server_logs.py
@@ -5,7 +5,7 @@ import asyncio
 
 from fastapi import APIRouter
 from fastapi.responses import StreamingResponse, Response
-from self_hosting_machinery.scripts import env
+from self_hosting_machinery.scripts import env  # REFACTORME
 
 
 class TabServerLogRouter(APIRouter):
